updates:
1) added class diagram
2) minor fixes in Commands subsystem
3) removed old files from repo

+++ src/draw_skeleton_video_3D/Points/BodyKeyPoint.cpp
@@ -15,7 +15,7 @@ void BodyKeyPoint::setConfidence (float confidence) {
 
 
 
-BodyKeyPoint::BodyKeyPoint (int x, int y, float confidence, Point * decorated) : Point(x, y, decorated) {
+BodyKeyPoint::BodyKeyPoint (float x, float y, float confidence, Point * decorated) : Point(x, y, decorated) {
     setConfidence(confidence);
 }
